Refactor player processing and logging
Include `StrategyPlayer.h` in `PlayerProccessor.cpp` to fix an incomplete type error. Enhance `APlayerProccessor::Initialize` to log `SInteractionsArray` contents. Change `ProcessPlayers` to return `void` and update player fitness with logging. Add logging for interaction values in `ReadFile`.

In `Resource.cpp`, update `SetTaked` to validate owner and call `ProcessPlayers`. In `Spawner.cpp`, set owner and initialize `APlayerProcessor` in `BeginPlay`, and add `ProcessPlayers` method.

Update `ProcessPlayers` signature in `PlayerProccessor.h` to `void`. Add `ProcessPlayers` method to `ASpawner` in `Spawner.h`. Change fitness handling in `StrategyPlayer.h` to use `float` instead of `int32`.

+++ Source/TacticInvasors/Public/StrategyPlayer.h
@@ -20,8 +20,8 @@ public:
 	bool InitializeAgent(ASpawner* InSpawner);
 	void SetAggresive(bool bAggresive) { bIsAgresive = bAggresive; }
 	bool IsAggresive() const { return bIsAgresive; }
-	void SetFitness(int32 InFitness) { Fitnees = InFitness; }
-	int32 GetFitness() const { return Fitnees; }
+	void SetFitness(float InFitness) { Fitnees = InFitness; }
+	float GetFitness() const { return Fitnees; }
 
 protected:
 	// Called when the game starts or when spawned
@@ -33,7 +33,7 @@ protected:
 	bool bIsAgresive = false;
 
 	UPROPERTY()
-	int32 Fitnees;
+	float Fitnees;
 
 	bool bHasTarget = false;
